+++ src/mockApi/authData.js
@@ -2,18 +2,196 @@ export const authData = {
   isAuthenticated: true,
   user: {
     id: "68b80a87-a122-4f73-8dc5-1b06ddbeeb96",
-    display: "MCBRIDE, CHRISTOPHER, E CAPT USAF ACC 74FS/CCV (Porkins)",
-    email: "christopher.mcbride.13@us.af.mil",
-    first: "Christopher",
-    middle: "Eugene",
-    last: "McBride",
-    rank: "Capt",
-    majcom: "ACC",
-    squadron: "74FS",
-    office: "CCV",
+    display: "SNUFFY, JOHN C AMN USAF AETC 315TRS/DOT",
+    email: "this.is.fake@us.af.mil",
+    first: "John",
+    middle: "C",
+    last: "Snuffy",
+    rank: "Amn",
+    majcom: "AETC",
+    squadron: "315TRS",
+    office: "DOT",
     role: "admin",
   },
   tests: [
+    {
+      id: '7455766d-f0e4-4fb7-b351-e7ac9e2e02bc',
+      mds: 'AFOQT - 1',
+      name: "SECTION 1: VERBAL ANALOGIES (25 QUESTIONS / 8 MINS)",
+      owner: '68b80a87-a122-4f73-8dc5-1b06ddbeeb96',
+      version: 1,
+      date: '11 May 2020',
+      seen: true,
+      questions: [
+        {
+          question: 'GLOVE is to HAND as SHOE is to',
+          options: ['SOCK', 'LEG', 'FOOT', 'TOE', 'HORSE'],
+          answer: 2,
+          reference: '',
+        },
+        {
+          question: 'AIRCRAFT is to FLY as BOAT is to',
+          options: ['STEER', 'SINK', 'SHIP', 'LAND', 'SAIL'],
+          answer: 4,
+          reference: '',
+        },
+        {
+          question: 'PLAN is to STRATEGY as FIGHT is to',
+          options: ['STRUGGLE', 'HIT', 'BLOW', 'BATTLE', 'CONQUER'],
+          answer: 3,
+          reference: '',
+        },
+        {
+          question: 'TRAVEL is to DESTINATION as WORK is to',
+          options: ['GOAL', 'OFFICE', 'LABOR', 'LEISURE', 'COMMUTE'],
+          answer: 0,
+          reference: '',
+        },
+        {
+          question: 'PLOW is to FIELD as RACQUET is to',
+          options: ['BALL', 'COURT', 'GANGSTER', 'HAND', 'JUDGE'],
+          answer: 1,
+          reference: '',
+        },
+        {
+          question: 'SALUTE is to RESPECT as HUG is to',
+          options: ['SQUEEZE', 'GESTURE', 'ARM', 'EMBRACE', 'AFFECTION'],
+          answer: 4,
+          reference: '',
+        },
+        {
+          question: 'GRAVITY is to PLANET as ODOR is to',
+          options: ['SKUNK', 'AROMA', 'NOSE', 'DEATH', 'SIGHT'],
+          answer: 0,
+          reference: '',
+        },
+        {
+          question: 'FISHING ROD is to HOOK as KNIFE is to',
+          options: ['BLADE', 'BULLET', 'CUT', 'STEAK', 'HANDLE'],
+          answer: 0,
+          reference: '',
+        },
+        {
+          question: 'BOOT is to HIKE as VISOR is to',
+          options: ['FACE', 'SUIT', 'MOTORCYCLE', 'HEAD', 'WELD'],
+          answer: 4,
+          reference: '',
+        },
+        {
+          question: 'PUSH is to SHOVE as CLIMB is to',
+          options: ['MOUNTAIN', 'FALL', 'WALK', 'LINGER', 'CLAMBER'],
+          answer: 4,
+          reference: '',
+        },
+        {
+          question: 'GROW is to MATURE as BLOOM is to',
+          options: ['ROSE', 'PETAL', 'FLOURISH', 'DECAY', 'BLOSSOM'],
+          answer: 4,
+          reference: '',
+        },
+        {
+          question: 'FUSELAGE is to BODY as',
+          options: ['TURRET is to TANK', 'LIFT is to DRAG', 'DIVING is to FALLING', 'WOOD is to OAK', 'COMPASS is to DIRECTION'],
+          answer: 0,
+          reference: '',
+        },
+        {
+          question: 'CLIMB is to RISE as DIVE is to',
+          options: ['SWIMMING POOL', 'ARC', 'FALL', 'DECREASE', 'SLIDE'],
+          answer: 2,
+          reference: '',
+        },
+        {
+          question: 'ROPE is to KNOT as THREAD is to',
+          options: ['SANDWICH', 'FABRIC', 'NEEDLE', 'SEWING', 'THIMBLE'],
+          answer: 1,
+          reference: '',
+        },
+        {
+          question: 'WATCH is to OBSERVE as SWITCH is to',
+          options: ['LIGHT FIXTURE', 'RAILROAD TRACK', 'HIDE', 'MOVE', 'REPLACE'],
+          answer: 4,
+          reference: '',
+        },
+        {
+          question: 'OBEY is to COMPLY as REPLY is to',
+          options: ['QUESTION', 'STATEMENT', 'ANSWER', 'REPOSE', 'REWIND'],
+          answer: 2,
+          reference: '',
+        },
+        {
+          question: 'HONOR is to HONESTY as TRUST is to',
+          options: ['DEPENDABILITY', 'TREASON', 'PROMPTNESS', 'INTELLIGENCE', 'EMPATHY'],
+          answer: 0,
+          reference: '',
+        },
+        {
+          question: 'SPEED is to DECELERATION as VELOCITY is to',
+          options: ['DISTANCE', 'THRUST', 'RAPIDITY', 'BRAKING', 'URGENCY'],
+          answer: 3,
+          reference: '',
+        },
+        {
+          question: 'COMPRESSION is to RAREFACTION as DENSE is to',
+          options: ['THIN', 'APOGEE', 'IGNORANT', 'COMPACTED', 'PROTECTION'],
+          answer: 0,
+          reference: '',
+        },
+        {
+          question: 'DELAY is to POSTPONE as PREVENT is to',
+          options: ['DETER', 'DISCOURAGE', 'OFFEND', 'DEFEND', 'SWITCH'],
+          answer: 0,
+          reference: '',
+        },
+        {
+          question: 'ADD is to SUBTRACT as GROW is to',
+          options: ['DECAY', 'MATURE', 'SHRINK', 'HEIGHT', 'DIVIDE'],
+          answer: 2,
+          reference: '',
+        },
+        {
+          question: 'LEFT is to PORT as RIGHT is to',
+          options: ['STARBOARD', 'DEPARTED', 'HARBOR', 'RED', 'GREEN'],
+          answer: 0,
+          reference: '',
+        },
+        {
+          question: 'ODOMETER is to DISTANCE as ALTIMETER',
+          options: ['ALTITUDE', 'DISTANCE', 'FUEL', 'INSTRUMENT', 'READOUT'],
+          answer: 0,
+          reference: '',
+        },
+        {
+          question: 'ELEVATOR is to HEIGHT as THRUST is to',
+          options: ['STAIRWAY', 'CLIMBING', 'ROCKET', 'ALTITUDE', 'FLIGHT'],
+          answer: 3,
+          reference: '',
+        },
+        {
+          question: 'LOOK is to INSPECT as ARRIVE is to',
+          options: ['DEPART', 'TRAVEL', 'LOCATE', 'ATTEND', 'NOTICE'],
+          answer: 3,
+          reference: '',
+        },
+      ]
+    },
+    {
+      id: '7455766d-f0e4-4fb7-b351-e7ac9e2e02bd',
+      mds: 'AFOQT - 1',
+      name: "SECTION 2: ARITHMETIC REASONING (25 QUESTIONS / 29 MINUTES)",
+      owner: '68b80a87-a122-4f73-8dc5-1b06ddbeeb96',
+      version: 1,
+      date: '11 May 2020',
+      seen: true,
+      questions: [
+        {
+          question: 'A theater contains x rows, with y seats in each row. How many total seats are there in the theater?',
+          options: ['x + y', 'x - y', 'xy', 'y - x', '2x + y'],
+          answer: 2,
+          reference: '',
+        },
+      ]
+    },
     {
       id: "e554c1b3-6713-4f06-a497-44800148a76b",
       mds: "A-10C",
@@ -21,6 +199,7 @@ export const authData = {
       owner: "68b80a87-a122-4f73-8dc5-1b06ddbeeb96",
       version: 3,
       date: "5 May 2020, 16:58 Zulu",
+      seen: true,
       questions: [
         {
           question: 'How many engines does the A-10C have?',
@@ -29,33 +208,33 @@ export const authData = {
           reference: 'T.O. A-10C-1'
         },
         {
-          question: 'How many engines does the A-10C have?',
+          question: 'How many wings does the A-10C have?',
           options: ['One', 'Two', 'Three', 'Four'],
           answer: 1,
           reference: 'T.O. A-10C-1'
         },
         {
-          question: 'How many engines does the A-10C have?',
+          question: 'How many APUs does the A-10C have?',
           options: ['One', 'Two', 'Three', 'Four'],
-          answer: 1,
+          answer: 0,
           reference: 'T.O. A-10C-1'
         },
         {
-          question: 'How many engines does the A-10C have?',
+          question: 'How many seats does the A-10C have?',
           options: ['One', 'Two', 'Three', 'Four'],
-          answer: 1,
+          answer: 0,
           reference: 'T.O. A-10C-1'
         },
         {
-          question: 'How many engines does the A-10C have?',
+          question: 'How many guns does the A-10C have?',
           options: ['One', 'Two', 'Three', 'Four'],
-          answer: 1,
+          answer: 0,
           reference: 'T.O. A-10C-1'
         },
         {
-          question: 'How many engines does the A-10C have?',
+          question: 'How many fuel tanks does the A-10C have?',
           options: ['One', 'Two', 'Three', 'Four'],
-          answer: 1,
+          answer: 3,
           reference: 'T.O. A-10C-1'
         }
       ]
@@ -67,6 +246,7 @@ export const authData = {
       owner: "68b80a87-a122-4f73-8dc5-1b06ddbeeb96",
       version: 1,
       date: "5 May 2020, 16:58 Zulu",
+      seen: true,
       questions: [
         {
           question: 'How many runways does Moody AFB have?',